bug.h: Drop filename from BUG/WARNING logs if building for TPL/SPL
On several platforms space is very tight when building for SPL or TPL.
To claw back a few bytes to be used for code remove the __FILE__ name
from the BUG() and WARN() type macros. Since those macros still print
the function name plus a line number this should not really affect
the ability to backtrace an actual BUG/WARN message to a specific
piece of code.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>